server/repository: add public API (get/list/search/lookup)
This commit also changes the dependency order so that repositories dependens on organizations, instead of the other way around.
Temporarily (during the migration), this means that the definition of OrganizationPrivateRead has been moved to endpoints.py
to avoid a circular dependency.

To get a list of all repositories in an organization, use the repository search.