server/dx: implement `request_correlation_id` in tasks logs
We had to move out the `on_job_start`/`on_job_end` approach.

Unfortunately, we don't have access to task arguments in this hook.

This prevents us to implement things like common arguments handling,
as we do for `request_correlation_id`.

To circumvent this limitation, we implement this behavior
through the `task_hooks` decorator.