clients: Better error handling in GitHub installation
Sometimes users report an error during installation. I suspect it might
be that GitHub redirects back immediately, but might not have fully
completed the installation (worker tasks). So our API seems to fail
without us capturing/handling it properly.

Informing the user to attempt to reload + throwing the error so Sentry
logs it for us to debug more closely.